Graphically illustrate the Floyd and Dijkstra algorithm
This is an application that is used to graphically illustrate the Floyd
and Dijkstra algorithm in step by step. User can use this application to
draw points and line then apply the selected algorithm on the graph.
Every single steps and change of parameter are display to help people
understand clearly the algorithm.